+++ src/handlers/roomHandler.js
@@ -1,0 +1,3 @@
+export const handleRoomRequest = (data, socket) => {
+    
+};
